View versioning
Adds versioning to Sync Gateway views.

Includes change to startup behaviour - database API will return '503 - Views Indexing' until views are available.
