+++ rest/bulk_api.go
@@ -235,7 +235,7 @@ func (h *handler) handleDump() error {
 	viewName := h.PathVar("view")
 	base.LogTo("HTTP", "Dump view %q", viewName)
 	opts := db.Body{"stale": false, "reduce": false}
-	result, err := h.db.Bucket.View(db.DesignDocSyncGateway, viewName, opts)
+	result, err := h.db.Bucket.View(db.DesignDocSyncGateway(), viewName, opts)
 	if err != nil {
 		return err
 	}
@@ -262,6 +262,7 @@ func (h *handler) handleDump() error {
 // HTTP handler for _repair
 func (h *handler) handleRepair() error {
 
+	// TODO: If repair is re-enabled, it may need to be modified to support xattrs
 	if true == true {
 		return errors.New("_repair endpoint disabled")
 	}
